python/mpc-server/requirements.txt
mcp[cli]
strawberry-graphql[fastapi]